services, dto y controller añadidos, de estos solo client a sido en parte implementado, modificacion de nombres y archivo dbCreate.sql para seguir la pauta, modificacion de tipo de algunos atributos para mantener coherencia con base de datos, tokens implementados y probados

+++ TBDBackendLab1/src/main/java/com/example/TBDBackendLab1/persistence/repository/ProductRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ProductRepository {
 
     ProductEntity addProduct(ProductEntity product);
-    ProductEntity getById(Long id);
-    List<ProductEntity> getByCategory(Long id);
+    ProductEntity getById(Integer id);
+    List<ProductEntity> getByCategory(Integer id);
 
 }
